Avancement Gestion (pas touché plz)

+++ DeleteEtudiant.php
@@ -29,6 +29,10 @@ if ( !empty($_POST)) {
         $q = $sql->prepare($dbh);
         $q->execute(array($identifiant_profil));
         $identifiant_profil = $q->fetchColumn();
+
+        $dbh = "SELECT identifiant_profil FROM profil";
+        if ($dbh ) 
+
         $dbh = "DELETE FROM permission WHERE id_profil = ?";    
         $q = $sql->prepare($dbh);
         $q->execute(array($identifiant_profil));
@@ -41,7 +45,8 @@ if ( !empty($_POST)) {
         $dbh = "DELETE FROM profil WHERE id_profil = ?";    
         $q = $sql->prepare($dbh);
         $q->execute(array($identifiant_profil));
-        header('Location: GestionEtudiant.php');
+        
+        header('Location: GestionEtudiant.php?error');
     }
 }
